⚡ Bolt: Optimize sigmoid memory allocation in _compute_yes_no_scores

+++ qwen3_embed/rerank/cross_encoder/qwen3_cross_encoder.py
@@ -203,7 +203,8 @@ class Qwen3CrossEncoder(OnnxTextCrossEncoder):
 
         # Fast sigmoid calculation on logit difference for 2-class classification (~10x faster)
         # ⚡ Bolt: Fast sigmoid using in-place operations to avoid array allocation overhead (~20% faster)
-        diff = yes_no_logits[:, 0] - yes_no_logits[:, 1]
+        # Type cast to float32 is required to prevent type errors during in-place mutation
+        diff = yes_no_logits[:, 0].astype(np.float32) - yes_no_logits[:, 1].astype(np.float32)
         with np.errstate(over="ignore"):
             np.exp(diff, out=diff)
             diff += 1.0
